lots of functions refactored out of the config
