+++ tests/test_algorithm.py
@@ -34,7 +34,7 @@ from pyrate.core.algorithm import (least_squares_covariance,
                                    factorise_integer,
                                    )
 
-from pyrate.core.config import parse_namelist
+from pyrate.configuration import parse_namelist
 from pyrate.core.shared import Ifg, convert_radians_to_mm
 from tests.common import small5_mock_ifgs, SML_TEST_TIF, UnitTestAdaptation
